requirements.txt
flask
flask-cors
transformers
torch
mecab-python3
requests
